\docType{package}
\name{gorthoR}
\alias{gorthoR}
\alias{gorthoR-package}
\title{Orthologs from gOrtholog database}
\description{
GorthoR provides tools to get orthologs, alignments and
trees from gOrtholog database, an ortholog database at
Genentech. More information about each function can be
found in its help documentation.
}
\details{
GetOrthologs - Find orthologs in other species, given
gene/protein IDs.

GetAllOrthologs - Get all orthologs in given species by
taxonomy IDs.

GetClusterId - Find orthologous cluster given gene/protein
IDs.

GetAlignment - Retrieve alignment of the given gene/protein
and its orthologs

GetTree - Retrieve tree of the given gene/protein and its
orthologs
}
\author{
Mark J Chen \email{chenj99@gene.com}
}
\references{
TBD
}

