\name{LoadGorthologData}
\alias{LoadGorthologData}
\title{Load gOrtholog data}
\usage{
LoadGorthologData(id.type = id.type)
}
\arguments{
  \item{type.id}{which type of IDs are used: Entrez Gene
  ID, Ensembl gene ID, Ensembl peptide ID}
}
\value{
loaded data
}
\description{
Load gOrtholog data
}
\examples{
LoadGorthologData()
}
\author{
Mark J Chen \email{chenj99@gene.com}
}

